MemberInitActivity.java 삭제 회원가입 한 액티비티로 합침(아직 스크롤뷰 안 함) 파이어베이스에 users 컬렉션 만들고 회원 가입 누르면 입력되도록 구현
